Change avoidCompressedOopsPenalty to actualy avoid compressed oops penalty
`-Xmx32G` does not actually lead to compressed references (ordinary
object pointers), contrary to what the function suggests.

Using openjdk64-11.0.2 the following behaviour can be observed:

* `-Xmx32G` - compressed oops not used
* `-Xmx31G` - compressed oops used
* `-Xmx32767M` - compressed oops not used
* `-Xmx32766M` - compressed oops used

So the actual threshold is ~ 32766 MB so just shy of 32GIB.

If `31G` is too much of a change compared to the previously returned
value of 32G, the memory threshold might have to be expressed in `M`
instead of `G`.