QUICK-224: Fit in the CircleCI memory
The default `medium` CircleCI agent has 4 GB of RAM.
Reduce memory footprint by disabling drilldowns.
It affects not only the CI IT, but the actual recommendations.
Drilldowns were not needed yet, so it's fine to disable them now.
If they become needed, we can reenable them and reevaluate the memory cost, which might be fixed by then.