JPERF-976: Do not test C52xlarge with 9 nodes
The test on 9 nodes keeps failing on "http://admin:admin@localhost:8080/rest/api/2/upgrade failed to get out of 503 status within PT18M". I tried to bump the timeouts, but it didn't help. The same problem used to affect instances with 10 nodes or more, so I decided to expand the scope of unwanted hardware configurations.